+++ app/layout.jsx
@@ -1,6 +1,6 @@
 import './globals.css';
-import Header from '../components/Header';
-import Footer from '@/components/Footer';
+import Header from '../components/layout/Header';
+import Footer from '@/components/layout/Footer';
 
 export default function RootLayout({ children }) {
     return (
