+++ components/client/ProductCarousel.jsx
